Rename airbrakes.py to context.py
